requirements.txt
aniso8601==1.1.0
boto3==1.4.0
botocore==1.4.58
Flask==0.11.1
future==0.15.2
jupyter==1.0.0
pip==9.0.1
PyGithub==1.27.1
python-dateutil==2.5.2
pytz==2016.4
redis==2.10.5
requests==2.9.1
rq==0.5.6
six==1.10.0
ipywidgets
authgen
rsyslog
